requirements.txt
pygame
moderngl
numpy
moviepy
diffusers
transformers
torch
imageio
Pillow
scipy